backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.10
joblib==1.4.2
nltk==3.9.1
pillow==11.0.0
psycopg2-binary==2.9.9
regex==2024.11.6
requests==2.32.3
sqlparse==0.5.1
tqdm==4.67.1
urllib3==2.2.3
